Fix base script
