+++ requirements.txt
@@ -1,0 +1,3 @@
+opencv-python==4.5.5.62
+Pillow==9.0.0
+PyAutoGUI==0.9.53
