+++ src/main.py
@@ -1,12 +1,14 @@
-from controller.worker import Worker
 import time
+
+from worker import Worker
+from control import Control
 
 def main():
     worker = Worker()
     while True:
         try:
             worker.execute()
-            time.sleep(10) # meaning that the worker will put all workers to work each 10 secs
+            time.sleep(10)
         except Exception:
             worker.restart()
